Adds logout page
